requirements.txt
customtkinter==5.2.2
Pillow==10.4.0
reportlab==4.2.2
